# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger warehouse API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1323",
    "paths": {
        "/auth/init": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/refresh-token": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/sign-in": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "parameters": [
                    {
                        "description": "json req body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Signin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/sign-out": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/stockmovementvehicle": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stockmovementvehicle"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "name": "createName",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "endReceivedGrossQuantity",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "endReceivedNetQuantity",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "endReceivedTareQuantity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "endReceivedTime",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "endSentGrossQuantity",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "endSentNetQuantity",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "endSentTareQuantity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "endSentTime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "fromWarehouseId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "preloads",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "productId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "relatedId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "",
                        "name": "sortField",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "startReceivedGrossQuantity",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "startReceivedNetQuantity",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "startReceivedTareQuantity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "startReceivedTime",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "startSentGrossQuantity",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "startSentNetQuantity",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "startSentTareQuantity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "startSentTime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "stockmovementId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "toWarehouseId",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "TRANSFER",
                            "IN",
                            "PURCHASE_ORDER",
                            "RETAIL"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "StockmovementTypeTransfer",
                            "StockmovementTypeIn",
                            "StockmovementTypePurchaseorder",
                            "StockmovementTypeRetail"
                        ],
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "vehicleId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/stockmovementvehicle/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stockmovementvehicle"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stockmovementvehicle"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.Signin": {
            "type": "object",
            "required": [
                "passwd",
                "username"
            ],
            "properties": {
                "passwd": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "example": "admindemo"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "payload": {
                    "type": "object"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}